requirements.txt
Django==1.7.1
Markdown==2.5.1
beautifulsoup4==4.3.2
dj-database-url==0.3.0
django-cors-headers==0.13
django-enumfield==1.1.1
django-extensions==1.4.6
django-filter==0.8
django-suit==0.2.11
djangorestframework==2.4.3
enum34==1.0.3
gunicorn==19.1.1
inflection==0.2.1
nltk==3.0.0
numpy==1.9.1
psycopg2==2.5.4
pycanlii==1.0.3b1
pywatson==0.1.0a2
requests==2.4.3
six==1.8.0
whitenoise==1.0.3
